requirements.txt
fastapi == 0.128.0
uvicorn == 0.40.0
motor == 3.7.1
pydantic == 2.12.5
python-dotenv == 1.2.1
pytest == 9.0.2
httpx == 0.28.1
